+++ src/fit.jl
@@ -1,9 +1,9 @@
 """   
-    fit(template::objective_function_template, data::Any, theta::Vector{Float64}; lower::Vector{Float64} = Vector{Float64}(), upper::Vector{Float64} = Vector{Float64}(), estimation_method::String = "M", br_method::String = "implicit_trace", optim_method = LBFGS(), optim_options = Optim.Options(), regularizer::Function = function regularizer(theta::Vector{Float64}, data::Any) Vector{Float64}() end)
+    fit(template::objective_function_template, data::Any, theta::Vector{Float64}; lower::Vector{Float64} = Vector{Float64}(), upper::Vector{Float64} = Vector{Float64}(), estimation_method::String = "M", br_method::String = "implicit_trace", optim_method = LBFGS(), optim_options = Optim.Options(), regularizer::Function = function regularizer(theta::Vector{Float64}, data::Any) Vector{Float64}() end) optim_arguments...)
 
 Fit an [`objective_function_template`](@ref) on `data` using M-estimation ([keyword argument](https://docs.julialang.org/en/v1/manual/functions/#Keyword-Arguments-1) `estimation_method = "M"`; default) or RBM-estimation (reduced-bias M estimation; [Kosmidis & Lunardon, 2020](http://arxiv.org/abs/2001.03786); [keyword argument](https://docs.julialang.org/en/v1/manual/functions/#Keyword-Arguments-1) `estimation_method = "RBM"`). Bias reduction is either through the maximization of the bias-reducing penalized objective in Kosmidis & Lunardon (2020) ([keyword argument](https://docs.julialang.org/en/v1/manual/functions/#Keyword-Arguments-1) `br_method = "implicit_trace"`; default) or by subtracting an estimate of the bias from the M-estimates ([keyword argument](https://docs.julialang.org/en/v1/manual/functions/#Keyword-Arguments-1) `br_method = "explicit_trace"`). The bias-reducing penalty is constructed internally using automatic differentiation (using the [ForwardDiff](https://github.com/JuliaDiff/ForwardDiff.jl) package), and the bias estimate using a combination of automatic differentiation (using the [ForwardDiff](https://github.com/JuliaDiff/ForwardDiff.jl) package) and numerical differentiation (using the [FiniteDiff](https://github.com/JuliaDiff/FiniteDiff.jl) package).
 
-The maximization of the objective or the penalized objective is done using the [**Optim**](https://github.com/JuliaNLSolvers/Optim.jl) package. Optimization methods and options can be supplied directly through the [keyword arguments](https://docs.julialang.org/en/v1/manual/functions/#Keyword-Arguments-1) `optim_method` and `optim_options`, respectively. `optim_options` expects an object of class `Optim.Options`. See the [Optim documentation](https://julianlsolvers.github.io/Optim.jl/stable/#user/config/#general-options) for more details on the available options.
+The maximization of the objective or the penalized objective is done using the [**Optim**](https://github.com/JuliaNLSolvers/Optim.jl) package. Optimization methods and options can be supplied directly through the [keyword arguments](https://docs.julialang.org/en/v1/manual/functions/#Keyword-Arguments-1) `optim_method` and `optim_options`, respectively. `optim_options` expects an object of class `Optim.Options`. Arguments (e.g. `autodiff = :forward`) can be passed directly to `Optim.optimize` through [keyword arguments](https://docs.julialang.org/en/v1/manual/functions/#Keyword-Arguments-1). See the [Optim documentation](https://julianlsolvers.github.io/Optim.jl/stable/#user/config/#general-options) for more details on the available options.
 
 An extra additive regularizer to either the objective or the penalized objective can be suplied via the [keyword argument](https://docs.julialang.org/en/v1/manual/functions/#Keyword-Arguments-1) `regularizer`, which must be a scalar-valued function of the parameters and the data; the default value will result in no regularization.
 
@@ -18,7 +18,8 @@ function fit(template::objective_function_template,
              br_method::String = "implicit_trace",
              optim_method = LBFGS(),
              optim_options = Optim.Options(),
-             regularizer::Function = function regularizer(theta::Vector{Float64}, data::Any) Vector{Float64}() end)
+             regularizer::Function = function regularizer(theta::Vector{Float64}, data::Any) Vector{Float64}() end,
+             optim_arguments...)
     if (estimation_method == "M")
         br = false
     elseif (estimation_method == "RBM")
@@ -46,9 +47,9 @@ function fit(template::objective_function_template,
     end
 
     if (length(lower) > 0) & (length(upper) > 0)
-        out = optimize(obj, lower, upper, theta, Fminbox(optim_method), optim_options)
+        out = optimize(obj, lower, upper, theta, Fminbox(optim_method), optim_options; optim_arguments...)
     else
-        out = optimize(obj, theta, optim_method, optim_options)
+        out = optimize(obj, theta, optim_method, optim_options; optim_arguments...)
     end
     if (estimation_method == "M")
         theta = out.minimizer